fix: put some default options to stack group
Those options are related to the `stack` group and should not be computed
for other groups since they may break others calls.

--debug is still a global option:

$ mergify --debug stack push

This does not work anymore:

$ mergify --token foobar stack push

Global stack options are to be passed like this:

$ mergify stack --token foobar push
